Organize prefilter functions
